Klocwork bug 114 * server-src/amindexd.c: Free their_feature_string. Klocwork bug 130 * restore-src/amrestore.c: Test maximum value for rst_flags->blocksize. Klocwork bug 294 * common-src/genversion.c: Check NULL result of malloc. Klocwork bug 294 * client-src/selfcheck.c: Check NULL result of fdopen. Klocwork bug 539, 542 * oldrecover-src/extract_list.c: Fix pfn2->next = fn2. * recover-src/extract_list.c   : Ditto Klocwork bug 268, 272, 543, 544 * oldrecover-src/extract_list.c: Check for cmd != NULL. * recover-src/extract_list.c   : Ditto Klocwork bug 510 * restore-src/amidxtaped.c: Check for argv != NULL. Klocwork bug 435 * changer-src/scsi-linux.c: Make buffer one larger. Klocwork bug 520 * changer-src/scsi-changer-driver.c (OpenDevice): Validate parameters. Klocwork bug 182 * changer-src/chg-scsi.c (clean_tape): Check usagetime != NULL.

+++ changer-src/chg-scsi.c
@@ -1,4 +1,4 @@
-static char rcsid[] = "$Id: chg-scsi.c,v 1.47 2006/07/05 19:57:39 martinea Exp $";
+static char rcsid[] = "$Id: chg-scsi.c,v 1.48 2006/07/06 11:57:28 martinea Exp $";
 /*
  * 
  *
@@ -1129,7 +1129,8 @@ clean_tape(
 
   if (drive_loaded(fd, drivenum))
     unload(fd, drivenum, cleancart);  
-  unlink(usagetime);
+  if (usagetime)
+    unlink(usagetime);
 }
 /* ----------------------------------------------------------------------*/
